+++ test/dummy/config/environments/development.rb
@@ -67,4 +67,6 @@ Rails.application.configure do
 
   # Uncomment if you wish to allow Action Cable access from any origin.
   # config.action_cable.disable_request_forgery_protection = true
+
+  config.hosts << URI.parse(ENV.fetch("PASSKIT_WEB_SERVICE_HOST")).host
 end