+++ test/dummy/db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[7.0].define(version: 2022_09_23_124956) do
+ActiveRecord::Schema[7.1].define(version: 2024_04_11_013721) do
   create_table "passkit_devices", force: :cascade do |t|
     t.string "identifier"
     t.string "push_token"
@@ -45,4 +45,20 @@ ActiveRecord::Schema[7.0].define(version: 2022_09_23_124956) do
     t.index ["passkit_device_id"], name: "index_passkit_registrations_on_passkit_device_id"
     t.index ["passkit_pass_id"], name: "index_passkit_registrations_on_passkit_pass_id"
   end
+
+  create_table "tickets", force: :cascade do |t|
+    t.string "name"
+    t.integer "user_id", null: false
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+    t.index ["user_id"], name: "index_tickets_on_user_id"
+  end
+
+  create_table "users", force: :cascade do |t|
+    t.string "name"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+  end
+
+  add_foreign_key "tickets", "users"
 end