+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## [0.7.0]
+- [#25](https://github.com/coorasse/passkit/pull/25): Change the label default color to black.
+
+## [0.6.1]
+
+- [#21](https://github.com/coorasse/passkit/pull/21): Support an ecryption key via `PASSKIT_URL_ENCRYPTION_KEY` environment variable.
+
+## [0.6.0]
+
+- [#20](https://github.com/coorasse/passkit/pull/20): Many new attributes added.
+
 ## [0.5.4]
 
 - Fix last-modified header format. Return it in RFC 2616 format.